Binary Search in java language
This is Binary Search Program in Java Language with basic algorithm on that i create static function in main class name is binarySearch with return type integer and have the four argunments one is array type of integer second is low, high and the key that we want to search return type integer on that function first we calculate mid of array because on binary search we start searching from mid of array so first we check mid of array if the number we check is on mi then return the array index if not then we check left side of array if not then we check right side of array and if number is not found we return -1 that key is not present in array so this function is work on recursion keep call itself until number is found or key is negative -1. on main function we call this function and provide integer array, last high and key that we want to search so here key is 4 that present on index 3.