Create Vasilije in Cacak.cpp
Aca and Milovan, two fellow competitive programmers, decided to give Vasilije a problem to test his skills.

Vasilije is given three positive integers: 𝑛 𝑘 and 𝑥 and he has to determine if he can choose 𝑘
 distinct integers between 1
 and 𝑛 such that their sum is equal to 𝑥

Since Vasilije is now in the weirdest city in Serbia where Aca and Milovan live, Cacak, the problem seems weird to him. So he needs your help with this problem.

For each test case output one line: "YES", if it is possible to choose 𝑘. distinct integers between 1 and 𝑛 such that their sum is equal to 𝑥 and "NO", if it isn't.